// frontend/app/src/project/directusTypes/types.ts
export type AboutUs = {
  date_created?: string;
  date_updated?: string;
  id: number;
  sort?: number;
  status: string;
  text_de?: string;
  user_created?: string & DirectusUsers;
  user_updated?: string & DirectusUsers;
};

export type Apartments = {
  available_from?: string;
  building?: number & Buildings;
  date_created?: string;
  date_updated?: string;
  family_friendly?: boolean;
  handicapped_accessible?: boolean;
  id: number;
  singleflat?: boolean;
  sort?: number;
  status: string;
  user_created?: string & DirectusUsers;
  user_updated?: string & DirectusUsers;
  washingmachines: string & Washingmachines[];
};

export type AppSettings = {
  avatars_enabled?: boolean;
  date_created?: string;
  date_updated?: string;
  enabledAvatarStyles?: string;
  id: number;
  initialAvatarStyle?: string;
  mealplaceholder_image?: string & DirectusFiles;
  status: string;
  user_created?: string & DirectusUsers;
  user_updated?: string & DirectusUsers;
  avatars_settings_group: string;
};

export type AppSettingsAccountBalance = {
  date_created?: string;
  date_updated?: string;
  enabled?: boolean;
  id: number;
  status: string;
  user_created?: string & DirectusUsers;
  user_updated?: string & DirectusUsers;
};

export type AppSettingsBuildings = {
  date_created?: string;
  date_updated?: string;
  enabled?: boolean;
  id: number;
  status: string;
  user_created?: string & DirectusUsers;
  user_updated?: string & DirectusUsers;
};

export type AppSettingsCourseTimetable = {
  date_created?: string;
  date_updated?: string;
  enabled?: boolean;
  id: number;
  status: string;
  user_created?: string & DirectusUsers;
  user_updated?: string & DirectusUsers;
};

export type AppSettingsFoods = {
  comments_type?: string;
  date_created?: string;
  date_updated?: string;
  enabled?: boolean;
  id: number;
  placeholder_image?: string & DirectusFiles;
  ratings_amount_display?: boolean;
  ratings_avg_display?: boolean;
  ratings_type?: string;
  status: string;
  user_created?: string & DirectusUsers;
  user_updated?: string & DirectusUsers;
};

export type AppSettingsHousing = {
  date_created?: string;
  date_updated?: string;
  enabled?: boolean;
  id: number;
  maps_enabled?: boolean;
  status: string;
  user_created?: string & DirectusUsers;
  user_updated?: string & DirectusUsers;
  translations: string & AppSettingsHousingTranslations[];
};

export type AppSettingsHousingTranslations = {
  app_settings_housing_id?: number & AppSettingsHousing;
  be_source_for_translations?: boolean;
  create_translations_for_all_languages?: boolean;
  description?: string;
  id: number;
  languages_code?: string & Languages;
  let_be_translated?: boolean;
};

export type AppSettingsNews = {
  date_created?: string;
  date_updated?: string;
  enabled?: boolean;
  id: number;
  status: string;
  user_created?: string & DirectusUsers;
};

export type AppSettingsNotifications = {
  android_enabled?: boolean;
  date_created?: string;
  date_updated?: string;
  email_enabled?: boolean;
  id: number;
  ios_enabled?: boolean;
  status: string;
  user_created?: string & DirectusUsers;
  user_updated?: string & DirectusUsers;
};

export type AppTranslations = {
  date_created?: string;
  date_updated?: string;
  id: string;
  sort?: number;
  status?: string;
  user_created?: string & DirectusUsers;
  user_updated?: string & DirectusUsers;
  translations: string & AppTranslationsTranslations[];
};

export type AppTranslationsTranslations = {
  app_translations_id?: string & AppTranslations;
  be_source_for_translations?: boolean;
  content?: string;
  create_translations_for_all_languages?: boolean;
  id: number;
  languages_code?: string & Languages;
  let_be_translated?: boolean;
  markdown?: string;
  translation: string;
};

export type AutoBackupSettings = {
  DB_CLIENT?: string;
  active?: boolean;
  backup_file_format?: string;
  backup_file_format_postfix?: string;
  backup_file_format_prefix?: string;
  backup_location_custom_path?: string;
  backup_location_folder_id?: string;
  backup_location_folder_name?: string;
  backup_location_type?: string;
  date_created?: string;
  date_updated?: string;
  id: number;
  latest_log?: string;
  sqlite3_db_filename?: string;
  state?: string;
  user_created?: string;
  user_updated?: string;
  backup_location_type_custom: string;
  backup_location_type_file_library: string;
  divide_to_database_type: string;
  divider-oql-hz: string;
  divider-qtbpzx: string;
  divider_to_backup_location: string;
  driver_specific_settings: string;
  general_settings: string;
  notice_finished: string;
  notice_please_configure_db_client: string;
  notice_please_configure_save_location: string;
  notice_please_setup_name_for_backupfile: string;
  notice_welcome: string;
  sqlite3_settings: string;
  visibility_active: string;
};

export type AutoTranslationSettings = {
  active?: boolean;
  auth_key?: string;
  extra?: string;
  id: number;
  informations?: string;
  limit?: number;
  percentage?: number;
  used?: number;
  valid_auth_key?: boolean;
  notice: string;
  usage: string;
  visible_for_valid_auth_key: string;
};

export type Buildings = {
  coord?: unknown;
  date_created?: string;
  date_updated?: string;
  id: number;
  image?: string & DirectusFiles;
  name?: string;
  sort?: number;
  status: string;
  user_created?: string & DirectusUsers;
  user_updated?: string & DirectusUsers;
  year_of_construction?: number;
  apartments: string & Apartments[];
  canteens: string & Canteens[];
  translations: string & BuildingsTranslations[];
};

export type BuildingsTranslations = {
  be_source_for_translations?: boolean;
  buildings_id?: number & Buildings;
  content?: string;
  create_translations_for_all_languages?: boolean;
  id: number;
  languages_code?: string & Languages;
  let_be_translated?: boolean;
  translation: string;
};

export type Businesshours = {
  date_created?: string;
  date_updated?: string;
  date_valid_from?: string;
  date_valid_till?: string;
  dayOfTheWeek?: string;
  id: number;
  sort?: number;
  status: string;
  time_end?: string;
  time_start?: string;
  user_created?: string & DirectusUsers;
  user_updated?: string & DirectusUsers;
};

export type Canteens = {
  building?: number & Buildings;
  date_created?: string;
  date_updated?: string;
  id: number;
  label?: string;
  sort?: number;
  status: string;
  user_created?: string & DirectusUsers;
  user_updated?: string & DirectusUsers;
  businesshours: string & CanteensBusinesshours[];
};

export type CanteensBusinesshours = {
  businesshours_id?: number & Businesshours;
  canteens_id?: number & Canteens;
  id: number;
};

export type ChatroomMessages = {
  chatroom?: number & Chatrooms;
  date_created?: string;
  date_updated?: string;
  id: number;
  sort?: number;
  status: string;
  textarea?: string;
  user_created?: string & DirectusUsers;
  user_updated?: string & DirectusUsers;
};

export type ChatroomTopics = {
  date_created?: string;
  date_updated?: string;
  id: number;
  sort?: number;
  status: string;
  user_created?: string & DirectusUsers;
  user_updated?: string & DirectusUsers;
  translations: string & ChatroomTopicsTranslations[];
};

export type ChatroomTopicsTranslations = {
  be_source_for_translations?: boolean;
  chatroom_topics_id?: number & ChatroomTopics;
  content?: string;
  create_translations_for_all_languages?: boolean;
  id: number;
  languages_code?: string & Languages;
  let_be_translated?: boolean;
};

export type Chatrooms = {
  date_created?: string;
  date_updated?: string;
  id: number;
  owner?: number & Profiles;
  sort?: number;
  status: string;
  topic?: number & ChatroomTopics;
  user_created?: string & DirectusUsers;
  user_updated?: string & DirectusUsers;
  messages: string & ChatroomMessages[];
};

export type Devices = {
  brand?: string;
  date_created?: string;
  date_updated?: string;
  display_fontscale?: string;
  display_height?: string;
  display_pixelratio?: string;
  display_scale?: string;
  display_width?: string;
  id: number;
  isAndroid?: boolean;
  isIOS?: boolean;
  isLandscape?: boolean;
  isSimulator?: boolean;
  isTablet?: boolean;
  isWeb?: boolean;
  platform?: string;
  profile?: number & Profiles;
  pushTokenObj?: unknown;
  sort?: number;
  status: string;
  systemVersion?: string;
  user_created?: string & DirectusUsers;
  user_updated?: string & DirectusUsers;
  display_group: string;
  system_group: string;
};

export type Flowhooks = {
  date_created?: string;
  date_updated?: string;
  id: number;
  last_parsing_date?: string;
  parse_foodoffers?: string;
  status: string;
  user_created?: string & DirectusUsers;
  user_updated?: string & DirectusUsers;
};

export type Foodoffers = {
  canteen?: number & Canteens;
  date?: string;
  date_created?: string;
  date_updated?: string;
  food?: string & Foods;
  id: number;
  price_employee?: number;
  price_guest?: number;
  sort?: number;
  status: string;
  user_created?: string & DirectusUsers;
  user_updated?: string & DirectusUsers;
  price_student?: number;
  markings: string & FoodoffersMarkings[];
};

export type FoodoffersMarkings = {
  foodoffers_id?: number & Foodoffers;
  id: number;
  markings_id?: number & Markings;
};
// ...
export type FoodsMarkings = {
  foods_id?: string & Foods;
  id: number;
  markings_id?: number & Markings;
};
// ...
export type CustomDirectusTypes = {
  about_us: AboutUs;
  apartments: Apartments;
  app_settings: AppSettings;
  app_settings_account_balance: AppSettingsAccountBalance;
  app_settings_buildings: AppSettingsBuildings;
  app_settings_course_timetable: AppSettingsCourseTimetable;
  app_settings_foods: AppSettingsFoods;
  app_settings_housing: AppSettingsHousing;
  app_settings_housing_translations: AppSettingsHousingTranslations;
  app_settings_news: AppSettingsNews;
  app_settings_notifications: AppSettingsNotifications;
  app_translations: AppTranslations;
  app_translations_translations: AppTranslationsTranslations;
  auto_backup_settings: AutoBackupSettings;
  auto_translation_settings: AutoTranslationSettings;
  buildings: Buildings;
  buildings_translations: BuildingsTranslations;
  businesshours: Businesshours;
  canteens: Canteens;
  canteens_businesshours: CanteensBusinesshours;
  chatroom_messages: ChatroomMessages;
  chatroom_topics: ChatroomTopics;
  chatroom_topics_translations: ChatroomTopicsTranslations;
  chatrooms: Chatrooms;
  devices: Devices;
  flowhooks: Flowhooks;
  foodoffers: Foodoffers;
  foodoffers_markings: FoodoffersMarkings;
  foods: Foods;
  foods_feedbacks: FoodsFeedbacks;
  foods_markings: FoodsMarkings;
  foods_translations: FoodsTranslations;
  image_overlays: ImageOverlays;
  image_overlays_translations: ImageOverlaysTranslations;
  languages: Languages;
  markings: Markings;
  markings_translations: MarkingsTranslations;
  news: News;
  news_translations: NewsTranslations;
  privacy_policy: PrivacyPolicy;
  profiles: Profiles;
  profiles_buildings_favorites: ProfilesBuildingsFavorites;
  profiles_buildings_last_visited: ProfilesBuildingsLastVisited;
  profiles_markings: ProfilesMarkings;
  washingmachines: Washingmachines;
  wikis: Wikis;
  wikis_translations: WikisTranslations;
  directus_activity: DirectusActivity;
  directus_collections: DirectusCollections;
  directus_fields: DirectusFields;
  directus_files: DirectusFiles;
  directus_folders: DirectusFolders;
  directus_migrations: DirectusMigrations;
  directus_permissions: DirectusPermissions;
  directus_presets: DirectusPresets;
  directus_relations: DirectusRelations;
  directus_revisions: DirectusRevisions;
  directus_roles: DirectusRoles;
  directus_sessions: DirectusSessions;
  directus_settings: DirectusSettings;
  directus_users: DirectusUsers;
  directus_webhooks: DirectusWebhooks;
  directus_dashboards: DirectusDashboards;
  directus_panels: DirectusPanels;
  directus_notifications: DirectusNotifications;
  directus_shares: DirectusShares;
  directus_flows: DirectusFlows;
  directus_operations: DirectusOperations;
};